requirements.txt
openai
langchain 
faiss-cpu
python-dotenv
pypdf
sentence-transformers
streamlit
chromadb
langchain-community
transformers
accelerate
groq 
streamlit 
openpyxl